fastapi-producer/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
kafka-python==2.0.2
uvicorn==0.15.0